Move audio output back into `main.rs`
This is mostly done because of some Rust lifetime issues and missing
`Send/Sync` traits of `rsoundio`. Callback registration via closures is
also non-trivial because they can't be passed by value because their
size is not known at compile time. This forces one to write some kind of
wrapper that is then passed into some `AudioIo` struct.
See: https://camjackson.net/post/rust-lang-how-to-pass-a-closure-into-a-trait-object