Add DSP module with envelope generator
The DSP module will contain any signal processing unit such as
Oscillators, Filters, Effects and the like.
The envelope generator uses a classic ADSR architecture where the attack
phase is logarithmic and the decay and release declines exponentially.
This mimics the charge and discharge cycles of the RC elements that
formed the envelope generators of analog synthesizers. Calculating the
envelope output is quite efficient because a one-pole filter is used to
generate the output.